requirements_container.txt
fastapi>=0.129.0
openai>=2.20.0
pydantic>=2.12.5
python-dotenv>=1.2.2
uvicorn>=0.40.0
